requirements.txt
et-xmlfile==1.1.0
lxml==4.9.4
openpyxl==3.1.2
python-docx==1.1.0
typing_extensions==4.9.0
